Exception handling

+++ AppointmentServiceNew/src/main/java/com/perficient/appointmentservice/service/DeleteAppointmentService.java
@@ -2,6 +2,7 @@ package com.perficient.appointmentservice.service;
 
 
 import com.perficient.appointmentservice.entity.Appointment;
+import com.perficient.appointmentservice.exception.AppointmentNotFoundException;
 import com.perficient.appointmentservice.repository.AppointmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +17,7 @@ public class DeleteAppointmentService {
 
     public void deleteAppointmentById(Integer aptId){
         Optional<Appointment> appointment = appointmentRepository.findById(aptId);
+        if (appointment.isEmpty()) throw new AppointmentNotFoundException("The Appointment doesn't exist");
         appointmentRepository.delete(appointment.get());
     }
 }
